Random seed updates for experiement running scripts

+++ assignment2/run_experiment.py
@@ -36,7 +36,7 @@ if __name__ == '__main__':
 
     seed = args.seed
     if seed is None:
-        seed = np.random.randint(0, (2 ** 32) - 1)
+        seed = np.random.randint(0, (2 ** 32) - 1, dtype='uint64')
         logger.info("Using seed {}".format(seed))
         np.random.seed(seed)
         rand.seed(seed)
